+++ src/main.jsx
@@ -21,6 +21,7 @@ import './v0514.css'
 import './v0516.css'
 import './v0517.css'
 import './v060.css'
+import './v070.css'
 
 createRoot(document.getElementById('root')).render(
   <React.StrictMode>